Feat: Add retrieving of event list with filtering

+++ main-service/src/main/java/ru/practicum/ewm/event/EventPatch.java
@@ -28,7 +28,7 @@ public record EventPatch(
         String description,
 
         @PositiveOrZero
-        Integer participantLimit,
+        Long participantLimit,
 
         Boolean paid,
         Boolean requestModeration,
